TheLuminaries_146_Autonomous Cyber Incident Response Agent/requirements.txt
# ============================================================================
# Core Web Framework & API
# ============================================================================
fastapi>=0.110,<1
uvicorn[standard]>=0.27
pydantic>=2.6
Jinja2>=3.1

# ============================================================================
# Database & ORM
# ============================================================================
SQLAlchemy>=2.0
psycopg[binary]>=3.1
pgvector>=0.3

# ============================================================================
# Configuration & Data Serialization
# ============================================================================
PyYAML>=6.0

# ============================================================================
# Machine Learning & Analytics (UEBA, Anomaly Detection, Feature Extraction)
# ============================================================================
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
tsfresh>=0.20
pyod>=2.0

# ============================================================================
# Graph/Correlation Utilities
# ============================================================================
networkx>=3.3
